requirements.txt
# Core NLP and ML libraries
transformers>=4.12.0
torch>=1.10.0
numpy>=1.20.0
scikit-learn>=0.24.0
spacy>=3.2.0
datasets>=1.11.0

# Document processing
pdfminer.six>=20200517
summa>=1.2.0

# Visualization
matplotlib>=3.4.0

# After installing the above requirements, run:
# python -m spacy download en_core_web_sm